Common/FunctionUtils.hpp Common/FunctionUtils.cpp * Added feature for constrainting the memory occupied by the descriptors rather than the number of descriptor files loaded.
Common/Makefile
* Added as comments the preprocessor directives that it uses.

Common/tests/Makefile
* Removed the preprocessor directives since it needs to be enabled where
the sources file is compiled.

VocabLearn/Makefile
* Added the preprocessor directives that it uses, they are included here
because VocabTree is a template class and hence is compiled on the
source file where it is used.

VocabLearn/VocabLearn.cpp
* Added some output status messages.

VocabLib/include/VocabTree.h
* Added commented code for implementing random descriptors selection, it
will be moved to a new application.

VocabLib/tests/Makefile
* Added as comments the preprocessor directives that it uses and removed
those which it doesn't.